requirements.txt
torch>=1.12.0
torchaudio
librosa
numpy
scikit-learn
matplotlib
seaborn
pandas
umap-learn
tqdm
scipy